adding lpop timeout float

+++ app/Record.py
@@ -120,7 +120,7 @@ class Record:
         self._waiters.append(future)
 
         try:
-            return await asyncio.wait_for(future, int(timeout)) if timeout and timeout != "0" else await future # pylint: disable=line-too-long
+            return await asyncio.wait_for(future, float(timeout)) if timeout and timeout != "0" else await future # pylint: disable=line-too-long
         except asyncio.TimeoutError:
             self._waiters.remove(future)
             return "$-1"
